requirements.txt
flask
python-dotenv
flask-smorest
flask-sqlalchemy
sqlalchemy
Flask-JWT-Extended
passlib
flask-migrate
gunicorn
flask-cors
requests
urllib3